[FEAT] 팔로우 객체 id 데이터 타입 변경
- 기존의 다른 id의 데이터 타입과 api에서 받아오는 데이터 타입이 Int 타입이어서 Long에서 Int로 변경

Fixed: TICO-431

+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/domain/model/Follow.kt
@@ -1,7 +1,7 @@
 package com.tico.pomorodo.domain.model
 
 data class Follow(
-    val followId: Long,
+    val followId: Int,
     val profileUrl: String? = null,
     val name: String,
     var isFollowing: Boolean
